irq-apple-fiq: FIQ code adjustments
In Apple M1 ARM64 SoCs, many interrupts are FIQs (fast IRQs) rather
than "real" IRQs. Those used to be handled by the IRQ driver, but it
makes more sense to have a separate driver as there are no
interdependencies. Like the IRQ driver, irq-apple-fiq relies on an
external vIPI layer since it doesn't provide a sufficient number of
IPIs.

A cpuhotplug state was added so CPUs don't come online with pending
FIQs, which would crash them as soon as the F flag in the PSTATE word
is cleared.

Signed-off-by: Pip Cet <pipcet@gmail.com>