t8103-j274.dts: properly reserve reserved memory, as mapped memory
Except for the framebuffer, of course, which simple-framebuffer
charmingly thinks is iomem.

We need this because page tables live in reserved memory sometimes and
the page table code assumes those are ordinary mapped addresses. Make
it so.

The defaults are merely there to indicate the intended order of the
entries.

Note that this requires fixes to the "loader", usually m1n1, which is
currently broken and fails to communicate real memory ranges to the
kernel.

Signed-off-by: Pip Cet <pipcet@gmail.com>