apple-dart: reuse locked page tables
The page table base registers (but not the page table contents
themselves) are locked by the bootloader on Apple M1 machines, so we
need to reuse the page tables that are in reserved memory.

Change the DART code not to die loudly when it encounters a locked
DART, to handle it correctly, to mark previously-mapped IOVAs as
reserved (this includes the frame buffer), and WARN if we get orders
to do anything we cannot do on a locked DART.

TODO: reserved page tables should never be freed, particularly the
pgd, which will be accessed even after a DART driver potentially
unloads.

TODO: unhardcode shifts and stuff during the page table load

TODO: @marcan mentioned he thought it might be useful to learn about
the existing mappings, and somehow use them. Figure out whether that
makes any sense.

TODO: maybe it does work for modular DART now

Signed-off-by: Pip Cet <pipcet@gmail.com>