apple-asc-mgmt: Apple M1 ASC management driver
The Apple M1 SoC offers multiple coprocessors accessed via a mailbox
protocol. In order to work with the lower layers, we need a "driver"
for the special endpoint 0. It's currently trivial, going through a
single message-response-message cycle as soon as other endpoints have
started up and the coprocessor sends the "powerack" message that may
or may not be about power.

My hunch is that this driver is going to end up doing more stuff...

TODO: headers

TODO: documentation

TODO: use proper phandles to access the rproc

Signed-off-by: Pip Cet <pipcet@gmail.com>