kvbox: abstraction for key-value mailboxes
This provides general support for devices which allow access to a
number of properties identified by keys. This includes the SMC
coprocessor on Apple M1 SoCs.

Unlike mailboxes, which have just a few channels, the SMC has more
than a thousand properties, and access to them doesn't have to be
exclusive.

TODO: move files to the right place, after figuring out where that is.

Signed-off-by: Pip Cet <pipcet@gmail.com>