gpio-kvbox: expose kvbox properties as GPIO pins
The Apple M1 SoC includes multiple coprocessors accessed via a mailbox
protocol. Some of them provide many properties, and some of those are
useful to access using the GPIO interface. Support that.

It's not clear whether the "gP12" pin is a GP(I)O pin or a pin for power
management only. The current code assumes it's GPIO.

TODO: should this be clock gates instead?

TODO: dt bindings, restructure

Signed-off-by: Pip Cet <pipcet@gmail.com>