apple-asc-smc: Apple SMC driver
The Apple M1 SoC includes an SMC which provides access to many
properties identified by fourcc keys. Support attaching to a work
endpoint and provide access to the properties as a kvbox.

TODO: move stuff to headers

TODO: code cleanup

TODO: figure out how to handle notifications. Do we really want to be
an IRQ chip?

Signed-off-by: Pip Cet <pipcet@gmail.com>