WIP: apple_m1_backlight: Apple M1 laptop backlight driver
The MacBook Pro laptop, based on the Apple M1 SoC, controls its
backlight through a combination of an SMC key (for turning it off and
on) and a firmware call to the DCP (for adjusting brightness).

Support that.

Note that there is a problem with potential instability in message
keys; however, it seems they are grouped in groups of 50 and the
endpoint we use is 352, which gives us some hope that it will be
stable across firmware versions.

Experimentally, power consumption varies linearly with the value
passed to the backlight control, so it's probably just a PWM's duty
cycle. We express the scale as linear in the hope that this is true.

TODO: boilerplate comments

TODO: header files

TODO: unhardcode constants

Signed-off-by: Pip Cet <pipcet@gmail.com>