mailbox: apple-bare-mailbox: single-channel mailbox for Apple M1
Restore comment. I once again believe this to be a good idea.

Yes, there's some overhead when doing funnelling mailbox->mailbox
drivers, but it's what the hardware wants. Performance shouldn't be
much of an issue.

The thing is that there are really three phases: ASC
initialization (during which nothing else must happen); EP
initialization; and EP operation. EP initialization often involves
allocating or acquiring shared memory, but EP operation is usually
trivial: just wait for a message and look at the shared buffer to see
what changed.

But EP initialization also involves, always, at least one message to
EP0.

We want the three phases to be as separate as possible, because we can
often share some of them but not all of them. The mailbox protocol
allows handover of channels, which is precisely what we want here. It
also allows describing the many layers of muxing in the device tree.

TODO: investigate whether it's worth it to queue more than one
message. This would mean sending txdone a bit later, since we only get
an interrupt when the queue is empty (but we could send txdone right
away in the case where we look at the mailbox to send another
message). I'm not going to do that unless there's a performance (or
functionality, of course) case to be made.

TODO: remove comment about @marcan. we've figured out this bug now.

TODO: move common stuff to headers

Signed-off-by: Pip Cet <pipcet@gmail.com>