WIP: apple-asc-syslog: Apple M1 ASC syslog endpoint driver
The Apple M1 SoC provides multiple coprocessors. Most of them offer a
syslog facility which sends simple messages to the main
system. Support that.

This is WIP because it doesn't work properly (the panic driver does)
and always prints the first message in the syslog buffer (rather than
the current one).

Note that it's a bad idea and hopefully unnecessary to treat the
syslog messages as anything but data for human consumption. The SMC,
for example, offers a notification mechanism which is more reliable
than parsing the syslog.

TODO: lots.

Signed-off-by: Pip Cet <pipcet@gmail.com>